BugsController: Removed BugUsersController and added actions in BugsController

+++ app/controllers/bugs_controller.rb
@@ -4,51 +4,93 @@ class BugsController < ApplicationController
   # frozen_string_literal: true
   before_action :check_authorization, only: %i[destroy edit update]
   before_action :check_user, only: %i[new show]
+  before_action :bug_params, only: %i[update]
   def index
     authorize Bug
+    dev_index if params[:project_id].present?
     project_id = UserProject.where(user_id: current_user.id).pluck(:project_id)
     @bugs = Bug.where(project_id: project_id)
   end
+
   def dev_index
-    #unassigned bugs
-    # ids = BugUser.where(user_id: current_user.id).pluck(:bug_id)
-    # @bugs = Bug.where.not(id: ids).where(project_id: params[:project_id])
-    @bugs=Bug.where.not(dev_id: nil).where(project_id: params[:project_id])
-    render 'unassigned'
+    unless check_project
+      @bugs = Bug.where.not(id: bug_found).where(project_id: params[:project_id])
+      render 'bug_users/unassigned'
+    end
   end
-  def show
+
+  def dev_show
     # authorize BugUser
     # ids = BugUser.where(user_id: current_user.id).pluck(:bug_id)
-    @bugs = Bug.where(dev_id: current_user.id, project_id: params[:id])
-
-    render 'assigned'
+    unless check_project
+      @bugs = Bug.where(id: bug_found, project_id: params[:id])
+      render 'bug_users/assigned'
+    end
   end
 
-  def show
-    authorize Bug
+  def dev_create
+    unless check_dev_create
+      @bug = Bug.find(params[:project_id])
+      dev = @bug.developer_ids unless @bug.developer_ids.nil?
+      dev.push(current_user.id)
+      if @bug.update(developer_ids: dev)
+        flash[:success] = 'Bug was successfully Assigned.'
+
+      else
+        flash[:error] = @bug.errors.full_messages.to_sentence
+      end
+
+      redirect_to project_bugs_path(@bug.project_id)
+    end
   end
 
-  def edit
-    authorize Bug
-  end
-
-  def destroy
-    authorize Bug
+  def dev_destroy
+    project_id = Bug.where(id: params[:id]).pluck(:project_id)
     @bug = Bug.find(params[:id])
+    dev = @bug.developer_ids
+    dev.delete(current_user.id.to_s)
 
-    if @bug.destroy
-      flash[:success] = 'Bug was successfully destroyed.'
+    if @bug.update(developer_ids: dev)
+      flash[:success] = 'Bug was successfully unassigned.'
 
     else
       flash[:error] = @bug.errors.full_messages.to_sentence
     end
-    redirect_to assigned_bugs_path
+
+    redirect_to assigned_bugs_path(project_id)
+  end
+
+  def show
+    authorize Bug
+    @bug = Bug.find(params[:id])
+  end
+
+  def edit
+    authorize Bug
+    @bug = Bug.find(params[:id])
+  end
+
+  def destroy
+    authorize Bug
+    if current_user.developer?
+      dev_destroy
+    else
+      @bug = Bug.find(params[:id])
+
+      if @bug.destroy
+        flash[:success] = 'Bug was successfully destroyed.'
+
+      else
+        flash[:error] = @bug.errors.full_messages.to_sentence
+      end
+      redirect_to assigned_bugs_path
+    end
   end
 
   def create
+    authorize Bug
     @bug = Bug.new(bug_params)
     check_user
-    authorize @bug, policy_class: BugPolicy
     respond_to do |format|
       if @bug.save
         format.html { redirect_to bug_url(@bug), flash: { success: 'Bug was successfully created.' } }
@@ -59,7 +101,8 @@ class BugsController < ApplicationController
   end
 
   def update
-    authorize @bug, policy_class: BugPolicy
+    authorize Bug
+    @bug = Bug.find(params[:id])
     respond_to do |format|
       if @bug.update(bug_params)
         format.html { redirect_to bug_url(@bug.id), flash: { success: 'Bug was successfully updated.' } }
@@ -70,16 +113,21 @@ class BugsController < ApplicationController
   end
 
   def new
-    @bug = Bug.new(project_id: params[:project_id])
-    authorize @bug, policy_class: BugPolicy
-    render 'bugs/new'
+    if current_user.developer?
+      dev_create
+    else
+      authorize Bug
+      @bug = Bug.new(project_id: params[:project_id])
+      render 'bugs/new'
+    end
   end
 
   def assigned
+    authorize Bug
     if current_user.qa?
-      @bugs = Bug.where(user_id: current_user.id)
+      @bugs = Bug.where(qa_id: current_user.id)
     elsif current_user.developer?
-      @bugs = User.find(current_user.id).bugs
+      params[:id] == 'list' ? @bugs = Bug.where(id: bug_found) : dev_show
     end
   end
 
@@ -91,7 +139,7 @@ class BugsController < ApplicationController
       params.require(:bug).permit(:piece_status)
     else
       params.require(:bug).permit(:id, :piece_status, :description, :title, :project_id, :deadline, :screenshot,
-                                  :piece_type).merge(user_id: current_user.id)
+                                  :piece_type).merge(qa_id: current_user.id)
     end
   end
 
@@ -104,13 +152,13 @@ class BugsController < ApplicationController
   end
 
   def check_qa
-    user_not_authorized if Bug.where(id: params[:id], user_id: current_user.id).take.nil?
     @bug = Bug.find(params[:id])
+    user_not_authorized if Bug.where(id: params[:id], qa_id: current_user.id).take.nil?
   end
 
   def check_dev
-    user_not_authorized if BugUser.where(bug_id: params[:id], user_id: current_user.id).take.nil?
     @bug = Bug.find(params[:id])
+    user_not_authorized unless @bug.developer_ids.include?(current_user.id.to_s)
   end
 
   def check_user
@@ -125,5 +173,26 @@ class BugsController < ApplicationController
   def check_show
     @bug = Bug.find(params[:id])
     user_not_authorized if UserProject.where(project_id: @bug.project_id, user_id: current_user.id).take.nil?
+  end
+
+  def bug_found
+    ids = []
+    @bugs = Bug.all
+
+    @bugs.sort.each do |bug|
+      ids.push(bug.id) if bug.developer_ids.include?(current_user.id.to_s)
+    end
+
+    ids
+  end
+
+  def check_dev_create
+    id = Bug.where(id: params[:project_id]).pluck(:project_id)
+    user_not_authorized if UserProject.where(project_id: id, user_id: current_user.id).take.nil?
+  end
+
+  def check_project
+    project_id = params[:project_id].nil? ? params[:id] : params[:project_id]
+    user_not_authorized if UserProject.where(project_id: project_id, user_id: current_user.id).take.nil?
   end
 end
